requirements.txt
djangorestframework
django 